% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gr.R
\name{annotate_gr}
\alias{annotate_gr}
\title{Annotate GR}
\usage{
annotate_gr(
  gr_query,
  gr_subject,
  cols_query = NULL,
  cols_subject = NULL,
  max_gap = -1L,
  min_overlap = 0L,
  as_boolean = FALSE,
  na_vals = "",
  collapse_as_string = FALSE,
  first_only = FALSE
)
}
\arguments{
\item{gr_query}{GRanges object to be annotated.}

\item{gr_subject}{GRanges object to be overlapped}

\item{cols_query}{Column(s) to be created in output GRanges object.}

\item{cols_subject}{Column(s) containing values to be inserted into query columns.}

\item{max_gap}{Maximum distance between features; defaults to 0bp. Must be set mutually exclusive to min_overlap.}

\item{min_overlap}{Minimum overlap between features; defaults to 1bp. Must be set mutually exclusive to max_gap.}

\item{as_boolean}{Should annotations be TRUE/FALSE (overlapped/not overlapped)? Defaults to FALSE. If TRUE, column values are ignored in favor of boolean values..}

\item{na_vals}{List of value(s) to annotate with in the event no overlaps are detected. Defaults to an empty string. Can be length 1 to apply to all columns, or one entry per column. Note that it must be a list if different data types are used in each column.}

\item{collapse_as_string}{Should annotations from multiple values be collapsed into a single semicolon-separated string? Defaults to FALSE, in which case these are returned as lists.}

\item{first_only}{Should only the first annotation be returned? Defaults to false; does not occur if <collapse_as_string> is TRUE.}
}
\description{
Given a GRanges object (query) and a second GRanges object to compare it to
(subject), identify all overlaps and annotate query by adding a new column
(cols_query) with values from a specified column on subject (cols_subject).
For instance, if cols_query is set as "peak_overlaps" and cols_subject
contains names of each peak in subject, function will return query with a
new column called "peak_overlaps" with the associated names from subject. If
no overlaps are found for a feature, it is annotated with the value of
<na_val>, which defaults to an empty string. If more than one overlap is
detected, all values are combined in a list, which can be collapsed into a
string separated by semicolons (collapse_as_string; can be slow) or reduced
to its first element (first_only). Function can also accept multiple query
and subject columns, provided these are the same length.

If a null value is provided to gr_query, columns of NA values will be added
to the original query and retruned.
}
