% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgzip_and_tabix.R
\name{bgzip_and_tabix}
\alias{bgzip_and_tabix}
\title{BGzip and Tabix}
\usage{
bgzip_and_tabix(
  file_in,
  silent = FALSE,
  bedtools_exe = "~/src/bedtools2/bin/bedtools",
  bgzip_exe = "/asclab/projects/shared/modules/samtools/samtools-1.2/bin/el7/bgzip",
  tabix_exe = "/asclab/projects/shared/modules/samtools/samtools-1.2/bin/el7/tabix"
)
}
\arguments{
\item{file_in}{Input file, must not be gzipped. If a .gz ending is present, the file will be presumed to not yet be compressed.}

\item{silent}{Boolean, should messages denoting steps be suppressed? Defaults to FALSE.}

\item{bedtools_exe}{BedTools executable.}

\item{bgzip_exe}{BGzip executable.}

\item{tabix_exe}{Tabix executable.}
}
\description{
Accepts an input bedGraph or BED file for bgzipping and tabix indexing.
Executable locations for Tabix, Bedtools, and BGzip are hard-coded.
}
