% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_to_distance.R
\name{bp_to_distance}
\alias{bp_to_distance}
\title{Trivia stats for basepair distances.}
\usage{
bp_to_distance(bp_value = sum(width(gr_seqs)))
}
\description{
Not particualrly useful, but good for trivia/context.
}
