% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_palette.R
\name{chrom_colors}
\alias{chrom_colors}
\title{Chromosome color palette}
\usage{
chrom_colors()
}
\description{
Returns a named vector of colors from ggsci::ucscgb useful for consistent
chromosome colors.
}
