% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_granges.R
\name{clip_granges}
\alias{clip_granges}
\title{Clip GRange}
\usage{
clip_granges(
  grange_in,
  grange_window,
  return_as_tibble = TRUE,
  replace_cols = FALSE,
  include_clipped_col = FALSE
)
}
\arguments{
\item{grange_in}{GenomicRange object(s) to be truncated. Can be a tibble, in which case it will be converted.}

\item{grange_window}{GenomicRange representing the window to be clipped to. Must be a single range.}

\item{return_as_tibble}{Should results be returned as a tibble? Defaults to FALSE.}

\item{replace_cols}{Should start and end values be changed to clipped values? Defaults to FALSE.}

\item{include_clipped_col}{Should a column be appended to metadata with the number of bases clipped? Defaults to FALSE.}
}
\description{
Trim features within a genomic range to fit within a specified window,
clipping feature starts/ends to fit within that region. If <replace_cols> is
TRUE, start and end columns are changed. Otherwise, start_adj and end_adj
columns are appended as metadata and the input GRange object is otherwise
unchanged. Can also be provided a tibble which will be converted to a GRange
assuming GenomicRanges::makeGRangesFromDataFrame() can interpret the columns
present.
}
