% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pngs.R
\name{combine_pngs}
\alias{combine_pngs}
\title{Combine PNGs}
\usage{
combine_pngs(p_file_list, n_row = 1, n_col)
}
\arguments{
\item{p_file_list}{List of PNG files to combine.}

\item{n_row}{Number of rows for the output image. Defaults to 1.}

\item{n_col}{Number of columns for the output image. If not provided, <n_col> is calculated based on the number of plots provided and <n_row>.}
}
\description{
Combine a list of PNG image filenames into a single PNG raster. Note that aspect ratios of input
files must be respected (no GGPlot2 rescaling takes place once a PNG is read as a raster). For
this reason, to combine two plots side-by-side in a 10" x 5" output plot, both input files should
first be saved as 5" x 5" PNG files.
}
