% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_bdg_table.R
\name{downsample_bdg_table}
\alias{downsample_bdg_table}
\title{Downsample BDG table.}
\usage{
downsample_bdg_table(
  table_in,
  bin_num = 2000,
  name_col = "name",
  bin_function = mean
)
}
\arguments{
\item{table_in}{Table of BedGraph values. Must have standard start/end/seqnames/score columns.}

\item{bin_num}{Number of bins to split coords into. Defaults to 2,000. If NULL or shorter than the actual range, the original table will be returned.}

\item{name_col}{Column containing names. Defaults to "name".}

\item{bin_function}{Function to combine bin values. Defaults to mean, but can be any function that takes in multiple values and returns a single number (sum, max,etc.).}
}
\description{
Given a table containing BedGraph format data, downsample genomic
coordinates such that approximately <bin_num> bins are included. Useful for
plotting (etc.) when BedGraphs of several megabases are to be shown on an
axis that will ultimately only span a few thousand pixels, both in terms of
storage and processing time. Binning is done separately for each chromosome
by splitting the the table by chromosome and running function on each
subset. Defaults to mean scores for each bin, but any meaningful function
can be provided via <bin_function>, "bin_function = max" for instance.
}
