% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factify.R
\name{factify}
\alias{factify}
\title{Factify}
\usage{
factify(table_in, max_unique_vals = 1000, max_char = 15, max_digits = 3)
}
\arguments{
\item{table_in}{Input table.}

\item{max_unique_vals}{Maximum unique values allowed in larger tables. Defaults to 1,000.}

\item{max_char}{Maximum number of characters allowed (all values in the column) in character columns. Defaults to 15.}

\item{max_digits}{Maximum number of digits allowed (all values in the column) in integer columns. Defaults to 3.}
}
\description{
Convert tibble columns to factors if they meet specific requirements. Factors
already present in the table are ignored, as are logical columns.

For character columns, factors must:
1) Have fewer than nrow() or <max_unique_values> unique values in the table.
2) Have fewer than <max_char> characters.

For numeric columns, factors must:
1) Be integers.
2) Have fewer than nrow() or <max_unique_values> unique values in the table.
3) Have fewer than <max_digits> digits.
}
