% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_shunt.R
\name{file_shunt}
\alias{file_shunt}
\title{File shunt}
\usage{
file_shunt(fileA, fileB, create_dir = FALSE, over_write = FALSE)
}
\arguments{
\item{fileA}{Origin file.}

\item{fileB}{Destination file.}

\item{create_dir}{Boolean; should the output directory be created if it is missing? Defaults to FALSE.}

\item{over_write}{Boolean; should an existing fileB be over-written? Defaults to FALSE.}
}
\description{
Given file names A and B, if file B does not exist copy file A to create B.
If file B does exist, ignore file A and return file B. Outputs file B after
creating / confirming it exists. Useful for creating local copies of a given
file.
}
