% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_and_exons.R
\name{genes_and_exons}
\alias{genes_and_exons}
\title{Get genes and exon data.}
\usage{
genes_and_exons(
  grange_win,
  return_exons = TRUE,
  clip_features = TRUE,
  cache_gene_tsv = NULL,
  cache_exon_tsv = NULL,
  overwrite_cache = FALSE,
  as_grange = FALSE
)
}
\arguments{
\item{grange_win}{GRange denoting region within which to find genes and exons.}

\item{return_exons}{Should exons be retrieved as well? Defaults to FALSE.}

\item{clip_features}{Should features be truncated to fit strictly within the window specified? Defaults to FALSE.}

\item{cache_gene_tsv}{TSV file in which to cache genes via clugPac::gtf_to_genes(). If not provided, no cache is created/read.}

\item{cache_exon_tsv}{TSV file in which to cache exons via clugPac::gtf_to_exons(). If not provided, no cache is created/read.}

\item{overwrite_cache}{Should existing cache TSVs be overwritten if found? Defaults to FALSE.}
}
\description{
Retrive gene and exon data from a GTF file and format as GRanges if
<as_granges> is TRUE. If no Grange region is provided, all genes are
retrieved, but exons are not.
}
