% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_add_bed_tiles.R
\name{genome_add_bed_tiles}
\alias{genome_add_bed_tiles}
\title{Genome: add BED tiles}
\usage{
genome_add_bed_tiles(
  plot_base_in,
  bed_files,
  bed_fills = "black",
  bed_colors = "black",
  bed_alphas = 0.3,
  bed_linetypes = "solid",
  bed_linewidths = 0.5
)
}
\arguments{
\item{plot_base_in}{Output stemming from <genome_plot()>.}

\item{bed_files}{List of BED files containing regions to highlight. Named BED files will be printed one-per-facet, unnamed files will be printed across all facets.}

\item{bed_fills}{List of fill values for ranges. Defaults to "black."}

\item{bed_colors}{List of color values for ranges. Defaults to "black."}

\item{bed_alphas}{List of alpha values for ranges. Defaults to 0.3.}

\item{bed_linetypes}{List of linetypes to apply to ranges. Defaults to "solid."}
}
\description{
Add a layer of tiles (rectangles) with an infinite y-axis range representing
start-end coordinates on the genome as stored in a BED file. Named files will
be separated into facets, while unnamed files will stretch over all facets.
}
