% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_add_bedgraph.R
\name{genome_add_bedgraph}
\alias{genome_add_bedgraph}
\title{Genome: add bedgraph}
\usage{
genome_add_bedgraph(
  plot_base_in,
  bdg_files,
  y_name = NULL,
  bdg_fills = "black",
  bdg_alphas = 1,
  bdg_multiplier = 1,
  facet_scales = "free_y",
  downsample_bins = 2000,
  coarse_downsample_lines = 1e+05
)
}
\arguments{
\item{plot_base_in}{Output stemming from <genome_plot()>.}

\item{bdg_files}{List of BedGraph files to be included. If named, each name will correspond to a single facet.}

\item{y_name}{Name to display on the y-axis. Defaults to "Pileup."}

\item{bdg_fills}{Fill value(s) to use for each bedGraph. Should be either length 1 for all facets or the same length as the number of bedGraph files. Defaults to "black".}

\item{bdg_alphas}{Alpha value(s) to use for each bedGraph. Defaults to 1.}

\item{bdg_multiplier}{If a multiplier value should be used to adjust bedGraph scores (such as with spike normalization), include the adjustment values here. Defaults to 1.}

\item{facet_scales}{Argument for ggplot2::facet_wrap(facet_scales). Defaults to "free_y".}
}
\description{
Add a layer of bedGraph signal to a genome_plot() list and return list along
with the new plot and additional tables for future layers. Function will also
use the clugPac::downsample_bdg_table() function to downsample output tables
and plots to keep approximately 2,000 bins on the X-axis. This number can be
adjusted using the <downsample_bins> argument, and to disable downsampling
all together use <downsample_bins) = NULL.
}
