% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_plot.R
\name{genome_plot}
\alias{genome_plot}
\title{Genome plot}
\usage{
genome_plot(gr_window, x_name = NULL, y_name = "Pileup")
}
\arguments{
\item{gr_window}{Single contiguous genomic range to display.}

\item{x_name}{Name to display on the x-axis. Defaults to NULL, in which case output of grange_desc() is used.}

\item{y_name}{Name to display on the y-axis. Defaults to "Pileup."}
}
\description{
Create a base plot for a genomic region. Produces a list containing GGplot
<plot> as well as a list of features for use with subsequent plot additions,
including the original grange <gr_win>,the x range, and a scale for
the X-axis.
}
