% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bed_saturation.R
\name{get_bed_saturation}
\alias{get_bed_saturation}
\title{Get BED saturation}
\usage{
get_bed_saturation(
  bed_files,
  name_out = "Saturation",
  cache_rds,
  iters = 10,
  extrap_points = 500,
  over_write = FALSE
)
}
\arguments{
\item{bed_files}{List of BED files to be analyzed. Assumes columns 1:3 are seqnames,start,end.}

\item{name_out}{Name column to be appended to each table. Defaults to "Saturation."}

\item{cache_rds}{RDS file to save results to and/or read cached results from. If not provided, no cache will be saved/loaded.}

\item{iters}{Number of randomized iterations to perform. Defaults to 10.}

\item{extrap_points}{Number of samples to extrapolate out in curve. Defaults to 500.}

\item{over_write}{Should existing <cache_rds> files be overwritten? Defaults to FALSE.}
}
\description{
Given a set of BED files, anticipate the saturation of annotated features by
unifying files step-wise in a random order, counting the number of unique 
features each time. Repeat the process for <iters> iterations, and use a 
saturation curve model (drc::drm(fct=22.5L())) to extrapolate peaks step-wise
up to <extrap_points> samples. Produce a plot summarizing this data, and 
store it as well as all tables and the saturation model to an RDS object, if
<cache_rds> is provided. If <cache_rds> is provided and that file exists,
read it directly (if <over_write> is FALSE).

Returns a list containing:
1) iterations: Mean and SD peaks added per sample.
2) predictions: Table of modeled mean peaks added per sample up to <extrap_points>.
3) fit: drc model and fit for saturation values.
4) details: Table of details for saturation analysis (steps, iterations, etc.)
5) plot: GGplot summarizing saturation.
}
