% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density_y.R
\name{get_density_y}
\alias{get_density_y}
\title{Get density Y-values.}
\usage{
get_density_y(vals_in, density_vals)
}
\arguments{
\item{vals_in}{Values to approximate.}

\item{density_vals}{All values used to measure the kernel density.}
}
\description{
Given a column of specific values, determine the approximate y-
position of each value in a kernel density plot (i.e. geom_density() in
ggplot2). Useful for adding labels to density plots.
}
