% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensembl_ids.R
\name{get_ensembl_ids}
\alias{get_ensembl_ids}
\title{Get Ensembl IDs}
\usage{
get_ensembl_ids(gene_symbols)
}
\description{
Retrieve Ensembl IDs from specific genes, if provided, otherwise return a
full table. IDs are retrieved from the same GTF files used by gtf_to_genes.
}
