% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exons.R
\name{get_exons}
\alias{get_exons}
\title{Get exons.}
\usage{
get_exons()
}
\description{
Retrieve a table of exons annotated with Ensembl IDs. Exon table was 
generated using TxDB and biomaRt via the script:

'/N/u/aclugston/resources/gene_annotations/generate_exon_table.R.
}
