% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_karyotypes.R
\name{get_karyotypes}
\alias{get_karyotypes}
\title{Get karyotypes.}
\usage{
get_karyotypes(grange_win = NULL, standard_chroms = TRUE)
}
\arguments{
\item{grange_in}{GRanges object denoting area to retrieve. Assumes a contiguous region; for non-contiguous regions use multiple calls.}
}
\description{
Get karyotype data from stored cytobands.txt file. Retreives either the full
genome or all bands within a specified genomic range. Includes "suggested"
colors in the "fill" column.
}
