% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_component_list.R
\name{get_plot_component_list}
\alias{get_plot_component_list}
\title{Get plot component list}
\usage{
get_plot_component_list(
  gtable_in,
  pattern,
  return_all = TRUE,
  drop_zero_grobs = TRUE
)
}
\arguments{
\item{gtable_in}{GTable object normally fed to get_plot_component.}

\item{pattern}{String pattern provided to get_plot_component used to select grobs ('guide-box-bottom', for instance).}

\item{return_all}{Should all matching grobs be returned (in a list)? Defaults to TRUE.}

\item{drop_zero_grobs}{Should zero (empty) grobs be ignored? Defaults to TRUE.}
}
\description{
Wrapper for cowplot::get_plot_component() which always returns a list and can
omit zero grobs (extremely reasonable yet somehow not standard behavior).
}
