% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seqsizes.R
\name{get_seqsizes}
\alias{get_seqsizes}
\title{Get seqsizes}
\usage{
get_seqsizes(coord_adjust = NULL, seqsize_file = NULL, as_granges = FALSE)
}
\arguments{
\item{coord_adjust}{GRanges coordinate to adjust; if provided, this coordinate will be returned with the appropriate adjustment made.}

\item{seqsize_file}{Defaults to NULL, in which case hg38_seqsizes.tsv is loaded from extdata. String of seqsizes TSV file to be used.}

\item{as_granges}{Boolean, should sizes be returned as a GRanges object containing one element per chromosome? Defaults to FALSE.}
}
\description{
Retrieve chromosome sizes for hg38 genome. By default, uses included
hg38_seqsizes.tsv file from extdata. If the argument <coord_adjust> is given
(as a genomic range object), the adjusted start value of this range will be
returned as an integer. Otherwise, a vector of chromosome size values is
returned. If <as_granges> is true, this vector is converted into a GRanges
object with seq size adjustments in a meta column.
}
