% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seqsizes_adj.R
\name{get_seqsizes_adj}
\alias{get_seqsizes_adj}
\title{Get seqsizes and adjust to end-to-end x-coordinates.}
\usage{
get_seqsizes_adj(seqsize_file = NULL)
}
\arguments{
\item{seqsize_file}{Filename of alternate seqsize TSV file to be used. Defaults to NULL, in which case hg38_seqsizes.tsv is loaded from extdata.}
}
\description{
Retrieve chromosome sizes for hg38 genome, then calculate adjusted X values
that allow chromosomes to be placed end-to-end (for instance, the "adjusted"
start of chr1 is 0, that of chr2 is the length of 1, that of chr3 is the sum
of the lengths of 1 and 2, and so on).
}
