% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tads.R
\name{get_tads}
\alias{get_tads}
\title{Get topologically associated domains (TADs)}
\usage{
get_tads(cell_types)
}
\arguments{
\item{cell_types}{A vector of cell types to be returned. If not provided,
all types are returned.}
}
\description{
Retrieve 3D genome browser TAD annotations. Downloaded on 22JUN08 from
http://3dgenome.fsm.northwestern.edu/downloads/hg38.TADs.zip.
}
