% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_themes.R
\name{gg_theme}
\alias{gg_theme}
\title{Clug themes}
\usage{
gg_theme(theme_name = "genome", ...)
}
\arguments{
\item{theme_name}{Name of theme of interest ("genome","scatter","barplot").}
}
\description{
Assorted GGplot2 themes, can be used similarly to 'theme_minimal()'.
Additional arguments to theme() can also be supplied.
}
