% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigBedToBed.R
\docType{data}
\name{gr_in}
\alias{gr_in}
\title{bigBedToBed}
\format{
An object of class \code{GRanges} of length 1.
}
\usage{
gr_in
}
\arguments{
\item{gr_in}{GRange to query.}

\item{bb_file}{Filename of bigBed format file.}

\item{exe_location}{Location of UCSC tools' bigBedToBed executable.}
}
\description{
Wrapper for UCSC tools; convert bigbed format file to BED format given an
input genomic range of interest.
}
\keyword{datasets}
