% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grange_desc.R
\name{grange_desc}
\alias{grange_desc}
\title{GRanges description}
\usage{
grange_desc(
  grange_in,
  force_unit = NULL,
  digits = 1,
  sep = "",
  append_ending = ""
)
}
\arguments{
\item{grange_in}{GRange to describe, should be of length one (if not, only the first will be described.)}

\item{force_unit}{Enforce a desired unit (bp, kb, MB) for width readout.}

\item{digits}{Number of digits to display, defaults to 1.}

\item{sep}{Separation between width value and units, defaults to "".}

\item{append_ending}{Insert a string between the width printout and the closing parenthesis; intended for bin size if relevant (for instance, "(1MB, 2kb bin size)").}
}
\description{
Given a GRanges object, returns a simple annotation in the format:

"Chr1: 1,000,000 - 1,100,000 (1MB)"
}
