% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_local_motifs.R
\name{grep_local_motifs}
\alias{grep_local_motifs}
\title{Grep Local Motifs}
\usage{
grep_local_motifs(gr_window, motif = "CTTAAG")
}
\arguments{
\item{gr_window}{GRanges object defining the contiguous region of the genome to search.}

\item{motif}{Motif in DNA alphabet to search for. Defaults to CTTAAG, which is the DLE1 binding motif for OGM.}
}
\description{
Given a genomic window of interest, search it for exact matches to a given
motif using the 'grep' function (i.e., not a standard motif searching
algorithm). Returns a GRanges object of motif locations.
}
