% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf_extract_features.R
\name{gtf_extract_features}
\alias{gtf_extract_features}
\title{GTF Extract Features}
\usage{
gtf_extract_features(
  gr_window,
  gtf_file = NULL,
  retain_attributes = NULL,
  do_not_process_attributes
)
}
\arguments{
\item{gr_window}{GRanges of regions to extract information from.}

\item{gtf_file}{Filename of Tabix-indexed GTF file to extract data from.}

\item{retain_attributes}{List of attributes to retain from column 9. For example, c("gene_name","transcript_id") will only keep gene names and transcript IDs while discarding the rest. Entries with one or both of these values missing will have NA values in their columns. No effect if <do_not_process_attributes> is TRUE.}

\item{do_not_process_attributes}{Should attribute column be returned unchanged (semi-colon separated list)? No default, in which case if more than 10000 features are returned processing is skipped with a warning.}
}
\description{
Given a GRange of genomic regions and a Tabix-indexed GTF file, read all 
features in the specified region(s) and return as a GRanges object. Any
attributes from column 9 are processed into a table with NA values, and 
these can be restricted using the <retain_attributes> argument.

If a large number of features is being retrieved, <do_not_process_attributes>
can be used to skip regex steps. By default, if T/F are not provided to this
argument then processing is skipped with a warning if more than 10k features
are returned from the scanning step. If processing is skipped, column 9 is 
returned as a semicolun-sparated list of attribute;"value" pairs.
}
