% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf_to_exons.R
\name{gtf_to_exons}
\alias{gtf_to_exons}
\title{GTF to exons}
\usage{
gtf_to_exons(
  granges_query,
  tabix_gtf,
  cache_tsv = NULL,
  overwrite_cache = FALSE,
  as_grange = TRUE
)
}
\arguments{
\item{granges_query}{GenomicRanges object of genomic regions to be searched.}

\item{tabix_gtf}{Tabix-indexed GTF file.}

\item{cache_tsv}{TSV file in which to cache exon results, if not provided no cache will be written/read.}

\item{overwrite_cache}{Should existing cache files be overwritten? Defaults to FALSE.}

\item{as_grange}{Should data be returned as a GRanges object? Defaults to TRUE, otherwise is returned as a tibble.}
}
\description{
Given a GTF file, extract all exons within region(s) specified by <granges_query>.
Returns a GRanges object. If no GTF file is provided, data is retrieved from
packaged "Homo_sapiens.GRCh38.104.chr.tabix.gtf.gz" file.
}
