% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isometrify_facets.R
\name{isometrify_table}
\alias{isometrify_table}
\title{Isometrify facets}
\usage{
isometrify_table(
  values_facet,
  values_x,
  values_y,
  x_rng_in = NULL,
  y_rng_in = NULL,
  x_rng_expand = 0.25,
  y_rng_expand = 0.25,
  include_demo = FALSE
)
}
\arguments{
\item{values_facet}{Vector of values containing facet information (i.e. categories to be spread in the Z-direction).}

\item{values_x}{Vector of x values.}

\item{values_y}{Vector of y values.}

\item{x_rng_in}{Vector of length 2 defining the x-range limits. If not provided, defaults to the range of <values_x>.}

\item{y_rng_in}{Vector of length 2 defining the y-range limits. If not provided, defaults to the range of <values_y>.}

\item{x_rng_expand}{Fraction of x-axis to expand when simulating the shift. All z-facet x-shifts will sum up to this value. Defaults to 0.25.}

\item{y_rng_expand}{Fraction of y-axis to expand when simulating the shift. All z-facet y-shifts will sum up to this value. Defaults to 0.25.}

\item{include_demo}{Should a demonstration isometric plot be provided? Defaults to FALSE.}
}
\description{
Given a table with very basic categorical faceting and x/y values, return a
a table columns for creating a pseudo-3D "isometric" view which
emulates three dimensions by shifting facet tracks in the x and y directions
and ordering geoms such geoms in the "foreground" will obstruct geoms in the
"background." See also: Diablo games.

Function is very brute-force; it returns columns that can be attached to an
original table using the "row_num" index column, x and y values, their
shifted minimums (x/y_shift), and their shifted values (x/y_shifted). The
depth column is generated using the factor levels of the <values_facet>
argument; if it is not a factor it will be made into one using the values in
the order submitted.
}
