% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_genecards.R
\name{link_genecards}
\alias{link_genecards}
\title{Link to GeneCards}
\usage{
link_genecards(gene_name = "six2", link_to_search = FALSE)
}
\arguments{
\item{gene_name}{Name or ID of gene to link to. Should work with ENS IDs, etc.}

\item{link_to_search}{Should link direct to a search page instead of the actual gene page? Defaults to FALSE.}
}
\description{
Link a given gene name to a GeneCards search: Basically pastes
the gene name/ID into a GeneCards URL and hopes for the best. Does not check
if links work, so should be reasonable when generating columns of a table.
}
