% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize_ogm_reads.R
\name{localize_ogm_reads}
\alias{localize_ogm_reads}
\title{Localize OGM Reads}
\usage{
localize_ogm_reads(gr_window, table_in, quiet = FALSE)
}
\arguments{
\item{gr_window}{GRanges describing the window to be searched.}

\item{table_in}{OGM table (as produced by localize_ogm_reads()) to be filtered and formatted.}

\item{quiet}{Boolean; should warnings about contigs with multiple locally aligned XMAPs be shown? Defaults to TRUE.}
}
\description{
To visualize OGM contigs relative to their local alignments, pseudo-alignment
values can be used to illustrate where the rest of the contig falls were it 
to be placed in a given genomic region. The localize_ogm_reads() function 
filters an OGM table to only include contigs with at least one XMAP alignment
to a genomic region of interest, and also appends columns to the output table
which include local pseudo-alignments. For instance, if a 1MB contig consists
of a locally-aligned 500MB XMAP and another 500MB XMAP aligned to a different
chromosome, the local XMAP column values will mirror its actual coordinates,
but the other XMAP will be assigned values that allow it to be drawn in the 
same location. In the event that multiple XMAPs align locally, a contig is 
returned for each along with a warning (unless 'quiet'=TRUE). Handling these 
may need to be done on a case-by-case basis. 

Added columns include:

local: T/F; is the tag part of a locally-aligned XMAP?

contig_xmap: Unique ID comprising contig_id_xmap_id.

contig_strand`: Direction of the entire contig *once aligned to the local 
XMAP*. In other words, if the local XMAP is in the minus direction the entire
contig is aligned in the minus direction.

local_contig_start: Local (pseudo) genomic position at which the full contig 
starts. Will be larger than local_contig_end if the contig is aligned in the 
3'-5' direction.

local_contig_end: Local (pseudo) genomic position at which the full contig 
ends.

local_xmap_start: Local (pseudo) genomic position at which a tag should be 
drawn. Should equal the local xmap_start/end values for locally aligned tags.
}
