% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bedgraph_grs.R
\name{merge_bedgraph_grs}
\alias{merge_bedgraph_grs}
\title{Merge BedGraph GRs}
\usage{
merge_bedgraph_grs(grA_in, grB_in, score_column = "score")
}
\arguments{
\item{grA_in}{GRanges A.}

\item{grB_in}{GRanges B.}

\item{score_column}{Score column name. Defaults to "score".}
}
\description{
Given two GRanges with bedGraph information, combine them and average their
score columns. Function uses pintersect() so the output range will typically
include additional features. For instance:

grA:   ====4====|====2======
grB:   =2=|========4========
grOut: =3=|==4==|=====3=====
}
