% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoKB_tables.R
\name{oncoKB_tables}
\alias{oncoKB_tables}
\title{OncoKB tables}
\usage{
oncoKB_tables(table_names = "drug_associations")
}
\arguments{
\item{table_names}{Vector of table names to retrieve.}
}
\description{
Retrieve OncoKB data table(s) as retrieved on 25JAN30. Tables include gene-
specific therapies (therapies), recognized alterations (variants), oncogenes
(oncogenes), drug/variant level descriptions (levels), and FDA-approved drug
descriptions ("drug_associations").
}
