% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bam_flags.R
\name{parse_bam_flags}
\alias{parse_bam_flags}
\title{Parse BAM flags}
\usage{
parse_bam_flags(bam_flag, filter_bits = NULL, as_boolean = FALSE)
}
\arguments{
\item{bam_flag}{Integer value to interpret.}

\item{filter_bits}{Restrict results to specific flags, for instance c(2,3) will only report "Properly aligned" or "Unmapped" flags. If NULL (default), all flags are returned concatenated with a semicolon separator.}

\item{as_boolean}{Boolean; should the presence of any flag be represented as T/F? Defaults to FALSE.}
}
\description{
Read BAM flags as integer values and interpret from binary. Useful for
describing a small number of reads in detail. Can return T/F for the presence
of flags, but if generating an index for a more complex filter use
flag_bam_reads(). If multiple flag values are submitted, a vector of the same
length will be returned. If no bam flags are submitted, a table summarizing
flag bits is returned.
}
\examples{
parse_bam_flags(return_flag_table=TRUE)
parse_bam_flags(1169)
parse_bam_flags(1169,filter_bits=11)
parse_bam_flags(1169,filter_bits=7,as_boolean=TRUE)
parse_bam_flags(1169,filter_bits=8,as_boolean=TRUE)

}
