% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ogm_reads.R
\name{parse_ogm_reads}
\alias{parse_ogm_reads}
\title{Parse OGM reads}
\usage{
parse_ogm_reads(fls_xmap, fls_cmap)
}
\arguments{
\item{fls_xmap}{Character vector or list of XMAP files to read. Can be named, and its names will take precedence over names of fls_cmap.}

\item{fls_cmap}{Character vector or list of CMAP files to read. Can be named.}
}
\description{
OGM data can be read from a combination of XMAP and CMAP files, which contain
contigs and their alignments to the genome called by Bionano and the
locations of fluorescent tags on each contig, respectively. The
`parse_ogm_reads()` function allows this to be done with pairs of XMAP/CMAP
files, and includes a name column which either contains the input XMAP
filename or, if multiple samples are read at the same time and at least one
of the two lists is named, the names assigned to the input XMAP/CMAP file
lists. This function reads pairs of files and compiles them into a single
table representing contigs (the actual DNA structure assembled from
overlapping reads), XMAPs (alignments assigned to each region of a given
contig), and tags (fluorescent tag locations). Since tags are only defined by
their position in a contig, this function calculates their putative *genomic*
location by determining which XMAP they fall in and measuring their position
based on its coordinates. Note that while XMAPs may overlap (there can be
ambiguity for alignments in some contigs), tag genomic locations are
calculated after assigning them to the highest-confidence XMAP. Also note
that tag locations may differ slightly as XMAP alignments are confidence
based.

The output table uses one row per tag with each tag's calculated genomic
location as the `seqnames` and `start` positions. Other fields are as follows:

strand: XMAP's purported alignment direction.

contig_id/xmap_id/site_id: ID values for the entire contig, each reported
genomic alignment *within* that contig (can be overlapping), and each
fluorescent tag location, respectively.

coverage: Total reads that comprise the contig which include each tag. I.e. a
coverage of 3 indicates that 3 of the molecules assembled into that contig
included this tag, so higher coverage suggests greater confidence.

contig_len: Length of the total contig in bp.

contig_start/end: Start/end positions of each XMAP alignment *within the
contig* in bp. I.e., an XMAP with contig_start of 2,000 starts 2kb from the
start of the contig. Note that contig_start is greater than contig_end in
minus-strand XMAPs.

xmap_start/end: *Genomic* start/end positions for each XMAP according to
Bionano.

xmap_conf: Confidence score of XMAP according to Bionano.

tag_contig_start: Location of tag *within the contig* in bp, i.e. a tag with
tag_contig_start of 2,000 is located 2kb from the start of the contig.

name: Name of the sample in question; defaults to the XMAP file name if only
one pair of XMAP/CMAP files are provided, but if a named list of files are
provided these names will be used.
}
