% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_coords.R
\name{pile_coords}
\alias{pile_coords}
\title{Pile coordinates}
\usage{
pile_coords(start_vals, end_vals, seqname_vals = NULL, min_gap = NULL)
}
\arguments{
\item{start_vals}{List or vector of start coordinates.}

\item{end_vals}{List or vector of end coordinates.}

\item{seqname_vals}{List or vector of chromosome names.}

\item{min_gap}{Minimum distance in basepairs between two items on the genome before they are incremented to another track. Defaults to NUll (1bp).}
}
\description{
Given a vector of start/end locations and optional chromosome names, return
an integer number corresponding to a stack position for each read (if read 2
overlaps read one, it gets a 2; if read 3 overlaps 1 and 2, it gets a three;
if read 4 doesn't overlap 1 through 3 it gets a 1, etc.). Vector names
correspond to index of the original vectors (1:n). All coordinates should be
the same length.
}
