% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_from_gtf.R
\name{plot_genes_from_gtf}
\alias{plot_genes_from_gtf}
\title{Plot Gene Track from GTF}
\usage{
plot_genes_from_gtf(
  genomic_region,
  gtf_file = NULL,
  max_support_level = 7,
  focus_genes = NULL,
  background_alpha = 0.25,
  intron_fill = "lightblue",
  exon_fill = "darkblue",
  label_color = "gray20",
  segment_linewidth = 0.5,
  text_size = 1.5,
  draw_arrows = TRUE,
  arrow_sweep = 0.005,
  arrow_size = 0.25,
  min_y_ax = 3,
  min_gap = 0.1,
  include_genes = NULL,
  exclude_genes = NULL,
  return_tables = FALSE
)
}
\arguments{
\item{genomic_region}{GRange denoting the region to be illustrated.}

\item{gtf_file}{GTF file to illustrate. If NULL (default), uses default GTF for gtf_extract_features().}

\item{max_support_level}{Max transcript support level allowed (1 == TSL1 == highest confidence). Defaults to 7, in which case all genes are shown.}

\item{focus_genes}{Gene names / EnsemblIDs to highlight by giving all other features a background alpha. Defaults to NULL, in which case all genes have an alpha of 1.}

\item{background_alpha}{Alpha value to use for background (non-<focus-genes>). Defaults to 0.25.}

\item{intron_fill}{Color to fill introns with. Defaults to lightblue.}

\item{exon_fill}{Color to fill exons with. Defaults to darkblue.}

\item{label_color}{Color to draw segments/errows/text for gene labels with. Defaults to gray20.}

\item{segment_linewidth}{Linewidths to draw segments/arrows with. Defaults to 0.25.}

\item{text_size}{Text size for gene names. Defaults to 1.5.}

\item{draw_arrows}{Boolean; should arrows be included? Defaults to TRUE.}

\item{arrow_sweep}{Distance to extend arrows upstream from TSS when drawing looped direction arrows. Units of X-range, defaults to 0.005 (0.5% of x-range).}

\item{arrow_size}{Size in cm for arrows to be drawn (arrow(length=unit(<arrow_size>,"cm"))). Defaults to 0.25.}

\item{min_y_ax}{Minimum y-axis height. Ensures that in cases where all annotations are on one level the plot area doesn't look too stretched. Defaults to 3, in which case no gene track will cover more than 30% of the panel's width.}

\item{min_gap}{Minimum gap between genes before they are moved to separate Y-tracks. Defaults to 0.1 (10% of the X-axis).}

\item{include_genes}{Names/EnsIDs of specific genes to include in the specified region. Useful for subsetting genes in the GTF file.}

\item{exclude_genes}{Names/EnsIDs of specific genes to exclude in the specified region. Useful for subsetting genes in the GTF file.}

\item{return_tables}{Boolean; should tables of data and label positions be returned instead of a plot? Defaults to FALSE.}

\item{feature_colors}{Colors to use for various feature types.}
}
\description{
Given a genomic range and a GTF file of feature annotations, plots a simple
gene track using gene annotations from gtf_extract_features(). If no gtf is
provided (gtf_file=NULL), uses default GTF file from gtf_extract_features().
Returns a GGplot2 object unless <return_tables> is TRUE, in which case a
list of tables for features, scales, and gene labels will be returned
instead.

Gene annotations include exons, and each gene represents the highest
confidence transcript for that gene/ensembl_id combination. In cases where
an exon's width is less than 1% of the x-range, that exon is drawn with a
vertical line to ensure all exons are still displayed.
}
