% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_pileup.R
\name{plot_genes_pileup}
\alias{plot_genes_pileup}
\title{Plot gene track pileup.}
\usage{
plot_genes_pileup(
  genomic_region,
  gr_genes_in = NULL,
  bin_num = 300,
  label_genes = NULL,
  label_size = 3,
  label_seed = NULL,
  kary_frac = 0.1
)
}
\arguments{
\item{genomic_region}{GRange denoting the region to be illustrated.}

\item{gr_genes_in}{GRanges object containing gene annotations. If not provided, gtf_to_genes() is used, but this can be time consuming: better to use gtf_to_genes() earlier and provide the output GR to this function directly.}

\item{bin_num}{Number of bins to split the range into. Defaults to 300.}

\item{label_genes}{Gene names to be labeled. Defaults to NULL, in which case no genes are labeled. Supersedes text_genes.}

\item{label_size}{Size of text in labels' font. Defaults to 3.}

\item{label_seed}{Seed number to use for ggrepel labels (used to maintain consistent label positions). Defaults to NULL.}
}
\description{
Given a genomic range, breaks the range into <bin_num> bins and plots the
number of genes per bin. Specific genes can be labeled using <label_genes>.
}
