% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_karyotype_strip.R
\name{plot_karyotype_strip}
\alias{plot_karyotype_strip}
\title{Plot karyotype track.}
\usage{
plot_karyotype_strip(
  genomic_region,
  window_region = "chrom",
  show_trace = TRUE,
  trace_line_color = "red",
  trace_point_size = 7,
  trace_line_width = 0.5,
  trace_alpha = 0,
  trace_fill = "red",
  trace_linetype = "solid",
  line_color_in = NA,
  line_widths_in = 0.5,
  label_size_in = 5,
  label_frac_in = 0.02
)
}
\arguments{
\item{genomic_region}{GRange denoting the region to be illustrated.}

\item{window_region}{Region to be illustrated overall; defualts to "chrom" (plots entire chromosome), and can be "genome" (plots entire genome). Other values cause only the selected region to be drawn.}

\item{trace_line_color}{Line color for box to draw illustrating the <genomic_region>.}

\item{trace_point_size}{Size of arrows on trace; defaults to 7.}

\item{trace_line_width}{Linewidth of trace; defaults to 0.5.}

\item{trace_alpha}{Alpha of trace; defaults to zero (empty red rectangle).}

\item{trace_fill}{Fill of trace; defaults to "red".}

\item{trace_linetype}{Linetype of trace; defaults to "solid".}

\item{line_color_in}{Color of lines around genomic bands; defaults to NA (no lines).}

\item{line_widths_in}{Width of lines around genomic bands; defaults to 0.5.}

\item{label_size_in}{Size of labels to draw; defaults to 5.}

\item{label_frac_in}{Fraction of x-axis below which bands are not labeled; defaults to 0.02, in which case bands which occupy less than 2% of the x-axis are drawn but not labeled.}
}
\description{
Plot genome-wide or chromosome-specific karytotypes. Plot labels/title are
included, but hidden by default using theme(). If <window_region> is set to
"chrom" or "genome", the strip will include the entire chromosome or genome
(respectively) and the <genomic_region> will be illustrated with a red box
and two red arrows.
}
