% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rna_genome.R
\name{plot_rna_genome}
\alias{plot_rna_genome}
\title{Plot genomic RNA}
\usage{
plot_rna_genome(
  gr_window_in,
  gr_genes_in,
  sample_names_in,
  sample_genes_in,
  sample_tpm_in,
  plot_metric = "TPM",
  z_scale = "log2",
  max_points = 20,
  point_size = 2,
  include_gene_track = "genes",
  widths_in = c(1, 12, 1),
  heights_in = c(1, 5, 5, 1),
  top_trace = TRUE,
  bottom_trace = FALSE
)
}
\arguments{
\item{gr_window_in}{GenomicRanges region to plot.}

\item{gr_genes_in}{GRanges of gene locations.}

\item{sample_names_in}{Vector of sample names for each TPM value.}

\item{sample_genes_in}{Vector of gene names for each TPM value.}

\item{plot_metric}{Value to plot. Defaults to TPM, can also be "ratio" in which case the ratio of expression among samples is plotted.}

\item{z_scale}{Scale for values. Defaults to log2, but can also be "log10."}

\item{max_points}{Maximum number of genes to display as scatter plots; defaults to 20.}

\item{point_size}{Point size for scatter plots.}

\item{include_gene_track}{Type of gene track to include. Defaults to "genes", but can also be "pileup" or "none."}

\item{table_tpm_in}{Vector of TPM values.}
}
\description{
Given a genomic range of gene locations (output by gtf_to_genes(), for
instance), and columns containing sample names, gene names, and TPM values,
plot RNA expression as scatter plots (if fewer than <max_points> genes are
present in the region) or tiles. Also outputs aligned gene tracks, by default.
}
