% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_split.R
\name{plot_split}
\alias{plot_split}
\title{Plot split.}
\usage{
plot_split(plot_in)
}
\arguments{
\item{plot_in}{GGPlot object to split.}
}
\description{
Split plot into basic top/right/bottom/left/main panels. Useful when 
plot_grid fails and when more specific controls are needed for plot component
heights/widths. Currently useful for broken cowplot::plot_grid() problem, but
very much a work in progress.
}
