% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transcripts.R
\name{plot_transcripts}
\alias{plot_transcripts}
\title{Plot transcripts}
\usage{
plot_transcripts(
  genomic_region,
  gtf_file = NULL,
  focus_genes = NULL,
  gene_text_size = 3,
  exon_text_size = 2,
  max_support_level = 7,
  max_transcripts = 7,
  label_exons = FALSE,
  return_tables = FALSE
)
}
\arguments{
\item{genomic_region}{Congiguous genomic region to illustrate.}

\item{gtf_file}{GTF file of feature annotations. If NULL, default GTF from gtf_extract_features() is used.}

\item{focus_genes}{Gene names/Ensembl IDs to highlight. Defaults to NULL, in which case all genes are given an alpha of 1.}

\item{gene_text_size}{Size of text on gene annotations. Defaults to 3.}

\item{exon_text_size}{Size of text on exon annotations. Defaults to 2.}

\item{max_support_level}{Maximum transcript support level (1 being highest confidence). Defaults to 7, which is any transcript support level.}

\item{max_transcripts}{Maximum number of transcripts to display for a given gene even if <max_support_level> filter would includ more. Defaults to 7, in which case the 7 highest confidence transcripts are shown.}

\item{label_exons}{Should exons be labeled with their exon number? Defaults to FALSE.}

\item{return_tables}{Should plotting be skipped and formatted tables be returned instead? Defaults to FALSE.}
}
\description{
Given a contiguous genomic region of interest and a GTF file, plot gene
transcripts and transcript features from that GTF file.
}
