% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyNumbers.R
\name{prettyNumbers}
\alias{prettyNumbers}
\title{Pretty numbers}
\usage{
prettyNumbers(num_in, force_unit = NULL, digs = 2)
}
\arguments{
\item{num_in}{Number to convert to a human-readable string.}

\item{force_unit}{Unit to convert numbers to, can be "T", "B", "M", "k" or "" for no units. Defaults to NULL, in which case output is decided automatically.}

\item{digs}{Digits to round numbers to. Defaults to 2.}
}
\description{
Given a number value, determine if it should be reported in thousands
(k, 1e3), millions (M, 1e6), billions (B, 1e9), or trillions (T, 1e12). Round
off digits to the <digs> place. Force a unit with <force unit>, which can be
any of "T","B","M","k", or "" (for no units). Values can be in string format,
provided that they are coercible to double (an error will be thrown
otherwise), and commas will be ignored such that "-223,442.05" can be read as
"-223k" by default.
}
