% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyTitle.R
\name{prettyTitle}
\alias{prettyTitle}
\title{prettyTitle}
\usage{
prettyTitle(
  str_in,
  sep = "_",
  caps = "first",
  preserve_levels = TRUE,
  preserve_ackronyms = TRUE
)
}
\arguments{
\item{argument_text}{String representing all arguments from a function. Can include arguments without a default and/or newlines.}
}
\description{
Given a string, function capitolizes either the first or every word (separated
by 'sep'). Runs recursively if given a list or vector, and if the input is a
factor with levels and 'preserve_levels' is TRUE, existing levels are also
changed and then reapplied. Otherwise, vectors are returned as character type.
}
