% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_text_file.R
\name{preview_text_file}
\alias{preview_text_file}
\title{Preview text file.}
\usage{
preview_text_file(file_name, lines_out = 12, suffix_line = "...")
}
\arguments{
\item{file_name}{Name of text file to preview.}

\item{lines_out}{Number of lines to preview; defaults to 12.}

\item{suffix_line}{Line to append to the tail of the output. Defaults to "...".}
}
\description{
Given a text file name (such as a script or README), print out 
the top <lines_out> lines followed by a final <suffix_line> using cat(). 
Useful for previewing text documents within markdowns.
}
