% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_grange.R
\name{randomize_grange}
\alias{randomize_grange}
\title{Randomize GRanges}
\usage{
randomize_grange(gr_in, max_sample = 10000, sample_num = 3, genome_in = "hg38")
}
\arguments{
\item{gr_in}{GRange to randomize.}

\item{max_sample}{Maximum number of features in <gr_in> to use per iteration.}

\item{sample_num}{Number of random iterations to perform. Defaults to 3.}

\item{genome_in}{String containing the genome that specified regions correspond to. Defaults to "hg38".}
}
\description{
Given an input GRange, randomly sample up to <max_sample> features within
this range and use regioneR's randomizeRegions() function to randomly
relocate them on the genome. Repeat this process <sample_num> times, and
return a single GRange with a column "set" labeling which iteration it was
produced in ("rand_3", for instance). Given that this function uses
regioneR, BSgenome.Hsapiens.UCSC.hg38 must be installed.
}
