% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bw_track.R
\name{read_bw_track}
\alias{read_bw_track}
\title{BigWig track}
\usage{
read_bw_track(
  bw_file,
  gr_window,
  bw_colnames = "score",
  new_colname = NULL,
  bin_num = 1000
)
}
\arguments{
\item{bw_file}{BigWig file of interest.}

\item{gr_window}{Region of interest in GRanges format.}

\item{bw_colnames}{Column names to retrieve in BigWig file; defaults to "score".}

\item{bin_num}{Number of bins to break data into, if desired. Defaults to 1,000, but can also be NULL for no binning.}
}
\description{
BigWig track
}
