% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cmap_file.R
\name{read_cmap_file}
\alias{read_cmap_file}
\title{Read CMAP file}
\usage{
read_cmap_file(cmap_fl)
}
\arguments{
\item{cmap_fl}{CMAP filename.}
}
\description{
Read a BioNano-formatted CMAP file containing OGM tag data (tag positions are
relative to contig position, not genome).
}
