% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastqc.R
\name{read_fastqc}
\alias{read_fastqc}
\title{Parse FastQC zips.}
\usage{
read_fastqc(zip_fastqc, stay_silent = TRUE)
}
\arguments{
\item{stay_silent}{Should scan() report how many lines were read? Defaults to FALSE.}

\item{zip_fastq}{FastQC Zip file(s) (with .zip ending) to be parsed.}
}
\description{
Given a FastQC output zip file, extract its content to a temp
directory and scan fastqc_data.txt. Remove temporary files, then parse data
file into constituent tables. If multiple Zip files are included, run
function recursively on each and append each into the same tables. Note that
tables with missing data are returned with the line "No data," and if run
recursively this may mean that some samples are not included in tables at
all ('No data' entries are omitted, so missing samples implicitly had no
data).

Note that  some FastQC versions appear to insert additional lines before a
given table's header row; these are ignored (for instance, the sequence
duplication levels table may have a "#Total deduplication percentage" tag).
}
