% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xmap_file.R
\name{read_xmap_file}
\alias{read_xmap_file}
\title{Read XMAP file}
\usage{
read_xmap_file(xmap_fl, as_granges = TRUE)
}
\arguments{
\item{xmap_fl}{XMAP filename.}

\item{as_grange}{Boolean; should the table be returned as a GRanges object? Defaults to TRUE.}
}
\description{
Read a BioNano-formatted XMAP file (file containing OGM 
alignment data) into a tibble containing XMAP and contig data. Can also 
return a GRanges object.
}
