% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_facets.R
\name{resize_facets}
\alias{resize_facets}
\title{Resize facets}
\usage{
resize_facets(plot_in, silent = FALSE)
}
\arguments{
\item{plot_in}{GGplot object.}

\item{silent}{Boolean, should warning messages be supressed? Defaults to FALSE.}
}
\description{
Given a GGplot with horizontal facets (typically barplots), build that plot
using ggplot_build in order to resize facets to fit data points. Has not
been fully implemented yet, so only works in cases of barplots or simple
scales, and does not work with vertical facets. Work in progress. Note that
this function should be a LAST step as things like themes stop working.
}
