% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to.R
\name{scale_to}
\alias{scale_to}
\title{Scale to}
\usage{
scale_to(values_in, min_range = -1, max_range = 1, min_val, max_val)
}
\arguments{
\item{values_in}{Vector of values to be scaled.}

\item{min_range}{Lower bound of range to scale, defaults to -1.}

\item{max_range}{Upper bound of range to scale, defaults to 1.}

\item{min_val}{Minimum value to arbitrarily tie the range to. If missing, the lowest number in <values_in> will be used.}

\item{max_val}{Maximum value to arbitrarily tie the range to. If missing, the highest number in <values_in> will be used.}
}
\description{
Scale a column of numbers between a minimum and maximum value.
}
