% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_bdg.R
\name{scan_bdg}
\alias{scan_bdg}
\title{Scan Tabix-indexed BedGraph file.}
\usage{
scan_bdg(bgz_files, name_vec, gr_regions, col_names)
}
\arguments{
\item{bgz_files}{List of filenames of tabix-indexed BDG files.}

\item{name_vec}{Vector of names for 'name' column. If missing, names can be retrieved from bgz_files, otherwise will use file base name. Must be the same length as bgz_files.}

\item{gr_regions}{GenomicRanges containing regions of interest.}

\item{col_names}{Column names of BedGraph file. Defaults to standard 5 columns (seq,start,end,score,name), but can be changed to include additional columns if necessary.}
}
\description{
Scan one or more BGzipped and Tabix-indexed files within regions specified
in <gr_regions> GRange object. Function is run recursively if more than one
bgzipped file is specified, and results are concatenated into a single
table. Uses Rsamtools::TabixFile for scanning and returns a tibble
}
