% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionInfoSorted.R
\name{sessionInfoSorted}
\alias{sessionInfoSorted}
\title{SessionInfoSorted}
\usage{
sessionInfoSorted()
}
\description{
Returns sessionInfo() output after sorting base, other, and non-attached package lists.
}
