% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_and_trim_gr.R
\name{subset_and_trim_gr}
\alias{subset_and_trim_gr}
\title{Subset and truncate GRanges}
\usage{
subset_and_trim_gr(
  gr_input,
  gr_window,
  note_truncated = FALSE,
  check_length = TRUE
)
}
\arguments{
\item{gr_input}{GRanges object to subset.}

\item{gr_window}{GRanges object containing one or more windows to <gr_input> to.}

\item{note_truncated}{Defaults to FALSE; if TRUE a start_offs and end_offs column will be added containing the number of basepairs trimmed from the start/end, respectively.}

\item{check_length}{Should function check that fewer than 1,000 query windows have been submitted? Defaults to TRUE.}
}
\description{
Given one GRanges object of features and another of length one
representing a region of interest, subset all features to only include those
that overlap the region of interest, and trim start/end locations to fit
within that window. If <note_truncated> is TRUE, append start/end_offset
columns with the number of basepairs removed from each object (i.e., if the
feature fell 100bp before the start of the window, add 100bp to the start
position and enter "100" in the start_offset column). Function can be run
recursively by submitting multiple gr_windows in a GRange, but if more than
1,000 entries are provided an error will display. This can be disabled by
setting <check_length>=FALSE.

An example usage would be:

subset_and_trim(gr_input = gene_list_grange, gr_window = regions_of_interest_gr)

...Which would return a GRanges object for each window containing the
truncated genes within each window.
}
