% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_columns.R
\name{swap_columns}
\alias{swap_columns}
\title{Swap Columns}
\usage{
swap_columns(tb_in, colA, colB, condition = NULL)
}
\arguments{
\item{tb_in}{Tibble to modify.}

\item{colA}{Column name (as a character vector) of first column.}

\item{colB}{Column name of second column.}

\item{condition}{Function of colA and colB values to determine if they should be swapped, or the name of a third column with boolean values. Defaults to NULL, i.e. swap all values.}
}
\description{
Given a table with at least two named columns, swap the values in columnA
(<colA>) with those of column B (<colB>) and vice versa, then return the
table. Swapping can be made conditional either by providing a function to
mapply which evaluates to TRUE/FALSE for each column value (for instance,
condition=function(x,y) x>y will swap values if column A is greater than
columnB). For more complex conditions or conditions which depend on other
columns in the table, a boolean column can be generated beforehand and its
name specified with a string.
}
\examples{
library(magrittr)
library(tibble)
library(dplyr)
library(tidyr)

# Dummy table.
tb <- tibble(idx=c(1:29)) \%>\%
 mutate(x=sample(1:1000,size=n()),
        y=sample(1:1000,size=n()))

swap_columns(tb,colA = "x",colB = "y")

# Swap columns if x < y
tb \%>\%
 swap_columns("x","y",function(x,y) x<y) \%>\%
 mutate(XgtY = x>y)

# Swap columns if y is even.
tb \%>\%
 swap_columns("x","y",function(x,y) y \%\% 2 == 0)

# Swap columns based on a third column.
tb \%>\%
 mutate(cond = sample(c(TRUE,FALSE),size=n(),replace=T)) \%>\%
 swap_columns("x","y","cond")

}
