% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_tibble.R
\name{trunc_tibble}
\alias{trunc_tibble}
\title{Truncate tibble.}
\usage{
trunc_tibble(
  table_in,
  n_row = 10,
  n_col = 10,
  row_suffix = "x...",
  col_suffix = "x...",
  sig_figs = 2
)
}
\arguments{
\item{table_in}{Table to be truncated.}

\item{n_row}{Number of rows to include.}

\item{n_col}{Number of columns to include.}

\item{row_suffix}{Suffix to describe truncated rows. Defaults to "x...", i.e. "23x...".}

\item{col_suffix}{Suffix to describe truncated columns in heading. Defaults to "y...", i.e. "23y...".}

\item{sig_figs}{Number of significant figures to round numbers to. Defaults to 2.}
}
\description{
Given a table input, print out <n_row> and <n_col> rows and columns,
respectively, with an added row/column of ellipses denoting how many have
been truncated. Useful for previewing large tables in an HTML.
}
