% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_distances.R
\name{tss_distances}
\alias{tss_distances}
\title{Transcription start site distances}
\usage{
tss_distances(gr_in, gr_genes_in, gene_id_col)
}
\arguments{
\item{gr_in}{GenomicRanges object of regions to measure.}

\item{gr_genes_in}{GenomicRanges object of gene annotations to measure against.}

\item{gene_id_col}{Name of metacolumn for <gr_genes_in> which will be used as gene IDs. Values will be returned as names of distance vector.}
}
\description{
Measure distances between a set of genomic regions and their nearest TSS. Negative
values indicate upstream respective of gene strand. Data is returned as a vector of integers. A
column name of gene IDs can be provided via <gene_id_col> which will be used to name the output
vector.
}
