% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectify.R
\name{vectify}
\alias{vectify}
\title{Vectify}
\usage{
vectify(table_in, value_col, name_col)
}
\arguments{
\item{table_in}{Input table.}

\item{value_col}{Name of column from which to derive values.}

\item{name_col}{Name of column from which to derive names.}
}
\description{
Starting with an input tibble, generate a named vector using one column for
names and another for values.
}
