% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oncogenes.R
\name{get_oncogenes}
\alias{get_oncogenes}
\title{Get oncogenes}
\usage{
get_oncogenes()
}
\description{
Retrieve the Bushman Lab group's compilation of cancer genes, downloaded
from http://www.bushmanlab.org/assets/doc/allOnco_June2021.tsv.

Website: http://www.bushmanlab.org/links/genelists
}
