% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_oxford.R
\name{oxford_collapse}
\alias{oxford_collapse}
\title{Oxford collapse}
\usage{
oxford_collapse(text_list)
}
\arguments{
\item{text_list}{List of character values to collapse.}
}
\description{
Collapse a list of things into a string with the correct Oxford
comma (Things, more things, and final things.)
}
