% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{plot_genes}
\alias{plot_genes}
\title{Plot gene track}
\usage{
plot_genes(
  genomic_region,
  gr_genes = NULL,
  label_genes = NULL,
  label_size = 3,
  label_seed = NULL,
  text_genes = NULL,
  text_size = 2,
  text_biotypes = c("protein_coding", "lncRNA"),
  text_seed = NULL
)
}
\arguments{
\item{genomic_region}{GRange denoting the region to be illustrated.}

\item{gr_genes}{GRanges object containing gene annotations. If not provided, gtf_to_genes() is used, but this can be time consuming: better to use gtf_to_genes() earlier and provide the output GR to this function directly.}

\item{label_genes}{Gene names to be labeled. Defaults to NULL, in which case no genes are labeled. Supersedes text_genes.}

\item{label_size}{Size of text in labels' font.}

\item{label_seed}{Seed number to use for ggrepel labels (used to maintain consistent label / text positions). Defaults to NULL.}

\item{text_genes}{Gene names to be labeled with text. Defaults to NULL, in which case no genes are labeled.}

\item{text_size}{Size of text font.}

\item{text_biotypes}{Vector of gene biotypes to be labeled with text. Defaults to "protein_coding" and "lncRNA." If "all" is listed, all biotypes will be labeled.}

\item{text_seed}{Seed number to use for ggrepel text (used to maintain consistent label / text positions). Defaults to NULL.}
}
\description{
Given a genomic range, plots a simple gene track using gene annotations from
gtf_to_genes(). Returns a GGplot2 object.
}
