% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{plot_genes}
\alias{plot_genes}
\title{Plot gene track}
\usage{
plot_genes(
  genomic_region,
  gr_genes = NULL,
  label_genes = NULL,
  label_size = 3,
  label_seed = NULL,
  focus_genes = NULL,
  background_alpha = 0.25,
  text_genes = NULL,
  text_biotypes = c("protein_coding", "lncRNA"),
  text_size = 2,
  text_bins = 10,
  text_nudge_y = 0.25,
  text_line_height = 0.2,
  text_nudge_x = 0.005,
  x_trace_genes = "label",
  x_trace_alpha = 0.3,
  arrow_genes = TRUE,
  gr_tads = NULL,
  tad_height = 1.5,
  tad_alpha = 0.1,
  tad_fill = "gray50",
  tad_color = NA,
  tad_linetype = "solid",
  tad_linewidth = 0.5,
  tad_y_frac = NULL,
  gr_eigs,
  return_tables = FALSE
)
}
\arguments{
\item{genomic_region}{GRange denoting the region to be illustrated.}

\item{gr_genes}{GRanges object containing gene annotations. If not provided, gtf_to_genes() is used, but this can be time consuming: better to use gtf_to_genes() earlier and provide the output GR to this function directly.}

\item{label_genes}{Gene names to be labeled. Defaults to NULL, in which case no genes are labeled. Supersedes text_genes.}

\item{label_size}{Size of text in labels' font. Defaults to 3.}

\item{label_seed}{Seed number to use for ggrepel labels (used to maintain consistent label positions). Defaults to NULL.}

\item{focus_genes}{Gene names to highlight (i.e. give an alpha of 1). If given, genes not in this list have their alpha reduced to <background_alpha>. Defaults to NULL, in which case all genes are alpha 1.}

\item{background_alpha}{Alpha value to use for background (non-<focus-genes>). Defaults to 0.25.}

\item{text_genes}{Gene names to be labeled with text. Defaults to NULL, in which case no genes are labeled. Genes must have non-empty gene name columns.}

\item{text_biotypes}{Vector of gene biotypes to be labeled with text. Defaults to "protein_coding" and "lncRNA." If "all" is listed, all biotypes will be labeled.}

\item{text_size}{Size of text font. Defaults to 2.}

\item{text_bins}{X-axis is split into <text_bins> "bins" and labels are stacked within each. Fewer bins result in higher stacks of labels, while lower bin numbers may result in more overlaps with labels/geometry. Defaults to 10.}

\item{text_nudge_y}{Distance above each text bin's highest gene annotation above which to start adding text labels in units of gene layers (one layer = one gene annotation). Defaults to 0.25.}

\item{text_line_height}{Height of text labels, useful to adjust in cases where larger text sizes are used causing them to overlap. Units of layers, defaults to 0.2.}

\item{text_nudge_x}{Distance to nudge labels in the negative x-direction when staggering stacked labels. Units of fractions of the X-axis, defaults to 0.005 (text is nudged 0.05 of the x-axis per stacked label).}

\item{x_trace_genes}{Gene names/IDs which should have traces drawn to the x-axis. Can be individual names/IDs, as well as "all", "focus" (foreground-only), and "label" (labeled genes only).}

\item{x_trace_alpha}{Opacity of x-axis traces. Defaults to 0.3, can be between zero and 1.}

\item{arrow_genes}{Should arrows be drawn for gene directions? Defaults to TRUE.}

\item{gr_tads}{GRanges of TADs to be drawn as diamonds in the background; defaults to NULL (none), and will be plotted in series separated by "name" column.}

\item{tad_height}{Y-position at which to center TAD diamonds. Defaults to 1.5 (centered on minimum gene level).}

\item{tad_alpha}{Opacity of TAD diamonds; defaults to 0.1.}

\item{tad_fill}{Fill of TAD diamonds; defaults to "gray50."}

\item{tad_color}{Color of TAD diamonds; defaults to NA (no outline).}

\item{tad_linetype}{Linetype of TADs; defaults to "solid."}

\item{tad_linewidth}{Linewidth of TADs; defaults to 0.5.}

\item{tad_y_frac}{Fraction of the y-axis to scale TADs to; defaults to depend on x-range.}

\item{return_tables}{Boolean; should tables of data and label positions be returned instead of a plot? Defaults to FALSE.}
}
\description{
Given a genomic range, plots a simple gene track using gene annotations from
gtf_to_genes(). Returns a GGplot2 object.
}
